pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bibtools"
version = "0.0.1"
dependencies = ["google-auth", "gspread", "pandas", "pyyaml", "requests"]

[project.optional-dependencies]
dev = [
  "ruff", # linting
]
